#10 Add asynchronous call to wait for the broker requests initialization to finish before initiating availability per day
When the availability files do not exists, we need to wait for the broker-related calls and respective processing to finish, in order to proceed with the processing of initiating the availability per day per clinic (for updating the markers availability). This is because we are relying on calculating the number of clinics based on the number of availability JSON files and when they are missing there is an error.